<🐛 fix>[client:layout]:{fix theme render not match}
