+++ client/app/layout.tsx
@@ -1,31 +1,52 @@
+import type { Metadata } from "next";
 import { ThemeProvider } from '@/components/theme-provider';
 import {NextIntlClientProvider} from 'next-intl';
 import {getMessages} from 'next-intl/server';
- 
+import localFont from "next/font/local";
+import Head from 'next/head';
+import "./globals.css";
+import "./globalicons.css";
+
+const geistSans = localFont({
+  src: "./fonts/GeistVF.woff",
+  variable: "--font-geist-sans",
+  weight: "100 900",
+});
+const geistMono = localFont({
+  src: "./fonts/GeistMonoVF.woff",
+  variable: "--font-geist-mono",
+  weight: "100 900",
+});
+
+export const metadata: Metadata = {
+  title: "Create Next App",
+  description: "Generated by create next app",
+};
+
 export default async function RootLayout({
   children,
   params: {locale}
-}: {
+
+}: Readonly<{
   children: React.ReactNode;
   params: {locale: string};
-}) {
+}>) {
   // Providing all messages to the client
   // side is the easiest way to get started
   const messages = await getMessages();
- 
   return (
     <html lang={locale}>
-      <body>
-      <ThemeProvider
-            attribute="class"
-            defaultTheme="system"
-            enableSystem
-            disableTransitionOnChange
-          >
-        <NextIntlClientProvider messages={messages}>
-          {children}
-        </NextIntlClientProvider>
-      </ThemeProvider>
+       <Head>
+        <link rel="preconnect" href="https://fonts.googleapis.com" />
+        <link rel="preconnect" href="https://fonts.gstatic.com" crossOrigin="anonymous" />
+        <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@100;300;400;500;700;900&display=swap" rel="stylesheet" />
+      </Head>
+      <body className={`${geistSans.variable} ${geistMono.variable} antialiased bg-background`}>
+        <ThemeProvider attribute="class" defaultTheme="system" enableSystem disableTransitionOnChange>
+          <NextIntlClientProvider messages={messages}>
+            {children}
+          </NextIntlClientProvider>
+        </ThemeProvider>
       </body>
     </html>
   );
